Migrate from Cypress 9 to 10
