+++ test/cypress/e2e/contact.spec.ts
@@ -1,4 +1,4 @@
-import { Product } from '../../../../data/types'
+import { Product } from '../../../data/types'
 
 describe('/#/contact', () => {
   beforeEach(() => {
