+++ test/cypress/e2e/dataErasure.spec.ts
