+++ test/cypress/e2e/deluxe.spec.ts
