+++ test/cypress/e2e/forgotPassword.spec.ts
