+++ test/cypress/e2e/privacyPolicy.spec.ts
