+++ test/cypress/e2e/redirect.spec.ts
