+++ test/cypress/e2e/register.spec.ts
