+++ test/cypress/e2e/scoreBoard.spec.ts
