+++ test/cypress/e2e/search.spec.ts
@@ -1,4 +1,4 @@
-import { Product } from '../../../../data/types'
+import { Product } from '../../../data/types'
 
 describe('/#/search', () => {
   beforeEach(() => {
