+++ test/cypress/e2e/tokenSale.spec.ts
